Extract content of share modal into its own component
